feat: add Simulation Controls toolbar
