+++ src/e3sm_compareview/components/toolbars.py
@@ -5,12 +5,17 @@ from trame.decorators import change
 from trame.widgets import client, html
 from trame.widgets import vuetify3 as v3
 
-from e3sm_compareview.comparison import COMPARISON_TYPES, COMPARISON_TYPE_LABELS
+from e3sm_compareview.comparison import (
+    COMPARISON_TYPES,
+    COMPARISON_TYPE_LABELS,
+    normalize_two_sim_target,
+)
 from e3sm_quickview.components.toolbars import Layout as Layout
 from e3sm_quickview.utils import js
 
 DENSITY = {
     "adjust-layout": "compact",
+    "simulation-controls": "compact",
     "comparison-controls": "compact",
     "adjust-databounds": "default",
     "select-slice-time": "default",
@@ -280,6 +285,332 @@ class ComparisonMode(v3.VToolbar):
                     classes="mr-2 text-none",
                     click="selected_columns = ['ctrl', 'test', 'diff', 'comp1', 'comp2']",
                 )
+
+
+class SimulationControls(v3.VToolbar):
+    def __init__(self):
+        super().__init__(**to_kwargs("simulation-controls"))
+
+        with self:
+            v3.VIcon("mdi-database-cog-outline", classes="pl-6 opacity-50")
+            with v3.VBtnToggle(
+                v_model=("comparison_mode", "multi-sim"),
+                mandatory=True,
+                density="compact",
+                divided=True,
+                classes="mx-3",
+            ):
+                v3.VBtn(
+                    "Two Sim",
+                    value="two-sim",
+                    variant="outlined",
+                    size="small",
+                    color="default",
+                    style=(
+                        "`border-width: ${comparison_mode === 'two-sim' ? '2px' : '1px'}; "
+                        "border-style: solid; "
+                        "border-color: ${comparison_mode === 'two-sim' ? 'rgb(var(--v-theme-primary))' : 'rgba(var(--v-border-color), var(--v-border-opacity))'}; "
+                        "background-color: ${comparison_mode === 'two-sim' ? 'rgba(var(--v-theme-primary), 0.14)' : 'transparent'}; "
+                        "color: ${comparison_mode === 'two-sim' ? '#000000' : 'rgba(var(--v-theme-on-surface), 0.7)'};`",
+                    ),
+                    classes=(
+                        "`text-none ${comparison_mode === 'two-sim' ? 'font-weight-bold' : 'text-medium-emphasis'}`",
+                    ),
+                )
+                v3.VBtn(
+                    "Multi Sim",
+                    value="multi-sim",
+                    variant="outlined",
+                    size="small",
+                    color="default",
+                    style=(
+                        "`border-width: ${comparison_mode === 'multi-sim' ? '2px' : '1px'}; "
+                        "border-style: solid; "
+                        "border-color: ${comparison_mode === 'multi-sim' ? 'rgb(var(--v-theme-primary))' : 'rgba(var(--v-border-color), var(--v-border-opacity))'}; "
+                        "background-color: ${comparison_mode === 'multi-sim' ? 'rgba(var(--v-theme-primary), 0.14)' : 'transparent'}; "
+                        "color: ${comparison_mode === 'multi-sim' ? '#000000' : 'rgba(var(--v-theme-on-surface), 0.7)'};`",
+                    ),
+                    classes=(
+                        "`text-none ${comparison_mode === 'multi-sim' ? 'font-weight-bold' : 'text-medium-emphasis'}`",
+                    ),
+                )
+
+            v3.VDivider(vertical=True, classes="mx-2")
+            with v3.VSelect(
+                v_model=("control_simulation_file", ""),
+                items=("simulation_configs", []),
+                item_title="label",
+                item_value="path",
+                label="Set control",
+                chips=True,
+                density="compact",
+                variant="solo",
+                hide_details=True,
+                disabled=("simulation_configs.length === 0",),
+                classes="mx-1",
+                style="min-width: 14rem; max-width: 18rem;",
+            ):
+                with v3.Template(v_slot_selection="{ item }"):
+                    with html.Div(classes="d-flex align-center py-1"):
+                        v3.VChip(
+                            "{{ item.raw.label || item.raw.path.split('/').pop() }}",
+                            size="small",
+                            color="primary",
+                            variant="outlined",
+                        )
+            with v3.VSelect(
+                v_if="comparison_mode === 'two-sim'",
+                v_model=("two_sim_test_simulation_file", ""),
+                items=(
+                    "simulation_configs.filter(sim => sim.path !== control_simulation_file)",
+                ),
+                item_title="label",
+                item_value="path",
+                label="Set test",
+                chips=True,
+                density="compact",
+                variant="solo",
+                hide_details=True,
+                disabled=(
+                    "simulation_configs.filter(sim => sim.path !== control_simulation_file).length === 0",
+                ),
+                classes="mx-1",
+                style="min-width: 14rem; max-width: 18rem;",
+            ):
+                with v3.Template(v_slot_selection="{ item }"):
+                    with html.Div(classes="d-flex align-center py-1"):
+                        v3.VChip(
+                            "{{ item.raw.label || item.raw.path.split('/').pop() }}",
+                            size="small",
+                            color="primary",
+                            variant="outlined",
+                        )
+
+            v3.VSpacer()
+
+            v3.VBtn(
+                "{{ (() => { const included = simulation_configs.filter(sim => sim.path === control_simulation_file || sim.include); return `Edit simulations (${included.length}/${simulation_configs.length || 0})`; })() }}",
+                size="small",
+                variant="outlined",
+                classes="text-none mr-3",
+                prepend_icon="mdi-pencil",
+                click="simulation_controls_dialog = true",
+            )
+
+            with v3.VDialog(
+                v_model=("simulation_controls_dialog", False),
+                max_width=720,
+                scrollable=True,
+            ):
+                with v3.VCard(style="max-height: 88vh;"):
+                    with v3.VToolbar(
+                        color="white",
+                        density="compact",
+                        classes="border-b-thin",
+                    ):
+                        v3.VIcon("mdi-database-cog-outline", classes="ml-4 mr-2")
+                        v3.VLabel("Simulation selection", classes="text-subtitle-2")
+                        v3.VSpacer()
+                        with v3.VTooltip():
+                            with v3.Template(v_slot_activator="{ props }"):
+                                html.Div(
+                                    "{{ simulation_configs.length }} loaded",
+                                    v_bind="props",
+                                    classes="text-caption mr-4",
+                                )
+                            html.Div(
+                                "{{ (() => { if (!simulation_configs.length) return 'Loaded simulations:\\nnone'; return `Loaded simulations:\\n${simulation_configs.map(sim => `${sim.label || sim.path.split('/').pop()}${sim.path === control_simulation_file ? ' (ctrl)' : ''}`).join('\\n')}`; })() }}",
+                                style="white-space: pre-line;",
+                            )
+                        v3.VBtn(
+                            icon="mdi-close",
+                            variant="text",
+                            size="small",
+                            classes="mr-2",
+                            click="simulation_controls_dialog = false",
+                        )
+                    with html.Div(
+                        v_if="simulation_configs.length === 0", classes="pa-4"
+                    ):
+                        html.Div(
+                            "Load simulation files first, then choose the control and comparison runs here.",
+                            classes="text-body-2 text-medium-emphasis",
+                        )
+
+                    with html.Div(
+                        v_else=True,
+                        classes="pa-3",
+                        style="max-height: calc(86vh - 64px); overflow-y: auto;",
+                    ):
+                        with html.Div(
+                            v_for="(entry, idx) in simulation_configs",
+                            key="`${entry.path}-card`",
+                            classes="simulation-entry-row pb-2 d-flex align-center",
+                        ):
+                            with html.Div(
+                                classes="d-flex flex-column align-center justify-center mr-2 ga-1",
+                            ):
+                                with v3.VTooltip(text="Move up"):
+                                    with v3.Template(v_slot_activator="{ props }"):
+                                        v3.VBtn(
+                                            v_bind="props",
+                                            icon="mdi-chevron-up",
+                                            size="small",
+                                            variant="outlined",
+                                            density="comfortable",
+                                            color="primary",
+                                            style="min-width: 34px; width: 34px; height: 34px;",
+                                            disabled=("idx === 0",),
+                                            click="""
+if (idx > 0) {
+  const nextConfigs = [...simulation_configs];
+  const [moved] = nextConfigs.splice(idx, 1);
+  nextConfigs.splice(idx - 1, 0, moved);
+  simulation_configs = nextConfigs;
+}
+""",
+                                        )
+                                with v3.VTooltip(text="Move down"):
+                                    with v3.Template(v_slot_activator="{ props }"):
+                                        v3.VBtn(
+                                            v_bind="props",
+                                            icon="mdi-chevron-down",
+                                            size="small",
+                                            variant="outlined",
+                                            density="comfortable",
+                                            color="primary",
+                                            style="min-width: 34px; width: 34px; height: 34px;",
+                                            disabled=(
+                                                "idx >= simulation_configs.length - 1",
+                                            ),
+                                            click="""
+if (idx < simulation_configs.length - 1) {
+  const nextConfigs = [...simulation_configs];
+  const [moved] = nextConfigs.splice(idx, 1);
+  nextConfigs.splice(idx + 1, 0, moved);
+  simulation_configs = nextConfigs;
+}
+""",
+                                        )
+                            with v3.VCard(
+                                variant="outlined",
+                                classes="flex-grow-1",
+                            ):
+                                with v3.VCardText(classes="pa-3"):
+                                    with v3.VRow(dense=True, classes="align-center"):
+                                        with v3.VCol(cols=12, md=6):
+                                            v3.VTextField(
+                                                model_value=("entry.label",),
+                                                update_modelValue="""
+simulation_configs = simulation_configs.map((sim) =>
+  sim.path === entry.path ? ({ ...sim, label: $event }) : sim
+);
+""",
+                                                label="Label",
+                                                density="compact",
+                                                variant="outlined",
+                                                hide_details=True,
+                                            )
+                                        with v3.VCol(cols=6, md=3):
+                                            with v3.VTooltip(
+                                                text=(
+                                                    "control_simulation_file === entry.path ? 'Current control simulation' : 'Set this simulation as control'",
+                                                ),
+                                            ):
+                                                with v3.Template(
+                                                    v_slot_activator="{ props }"
+                                                ):
+                                                    v3.VBtn(
+                                                        v_bind="props",
+                                                        text=(
+                                                            "control_simulation_file === entry.path ? 'Control' : 'Set control'",
+                                                        ),
+                                                        variant="outlined",
+                                                        color=(
+                                                            "control_simulation_file === entry.path ? 'primary' : 'default'",
+                                                        ),
+                                                        classes=(
+                                                            "`text-none w-100 ${control_simulation_file === entry.path ? '' : 'text-medium-emphasis'}`",
+                                                        ),
+                                                        style="min-width: 112px;",
+                                                        size="small",
+                                                        click=(
+                                                            self._on_control_selected,
+                                                            "[entry.path]",
+                                                        ),
+                                                    )
+                                        with v3.VCol(cols=6, md=3):
+                                            with v3.Template(
+                                                v_if="comparison_mode === 'multi-sim'"
+                                            ):
+                                                with v3.VTooltip(
+                                                    text="Toggle simulation inclusion",
+                                                ):
+                                                    with v3.Template(
+                                                        v_slot_activator="{ props }"
+                                                    ):
+                                                        v3.VCheckbox(
+                                                            v_bind="props",
+                                                            model_value=(
+                                                                "control_simulation_file === entry.path ? true : entry.include",
+                                                            ),
+                                                            update_modelValue="""
+simulation_configs = simulation_configs.map((sim) =>
+  sim.path === entry.path ? ({ ...sim, include: !!$event }) : sim
+);
+""",
+                                                            label="Include",
+                                                            density="compact",
+                                                            hide_details=True,
+                                                            disabled=(
+                                                                "control_simulation_file === entry.path",
+                                                            ),
+                                                        )
+                                            with v3.Template(v_else=True):
+                                                with v3.VTooltip(
+                                                    text=(
+                                                        "control_simulation_file === entry.path ? 'Control run cannot also be test' : (two_sim_test_simulation_file === entry.path ? 'Current test simulation' : 'Set this simulation as test')",
+                                                    ),
+                                                ):
+                                                    with v3.Template(
+                                                        v_slot_activator="{ props }"
+                                                    ):
+                                                        v3.VBtn(
+                                                            v_bind="props",
+                                                            text=(
+                                                                "two_sim_test_simulation_file === entry.path ? 'Test' : 'Set test'",
+                                                            ),
+                                                            variant="outlined",
+                                                            color=(
+                                                                "two_sim_test_simulation_file === entry.path ? 'primary' : 'default'",
+                                                            ),
+                                                            classes=(
+                                                                "`text-none w-100 ${two_sim_test_simulation_file === entry.path ? '' : 'text-medium-emphasis'}`",
+                                                            ),
+                                                            style="min-width: 112px;",
+                                                            size="small",
+                                                            disabled=(
+                                                                "control_simulation_file === entry.path",
+                                                            ),
+                                                            click="two_sim_test_simulation_file = entry.path",
+                                                        )
+                                    html.Div(
+                                        "{{ entry.path }}",
+                                        classes="text-caption text-medium-emphasis mt-2",
+                                        style="overflow: hidden; text-overflow: ellipsis; white-space: nowrap; direction: rtl; text-align: left;",
+                                        title=("entry.path",),
+                                    )
+
+    def _on_control_selected(self, control_path, **_):
+        self.state.control_simulation_file = control_path
+        if self.state.comparison_mode != "two-sim":
+            return
+
+        self.state.two_sim_test_simulation_file = normalize_two_sim_target(
+            self.state.simulation_configs,
+            control_path,
+            self.state.two_sim_test_simulation_file,
+        )
 
 
 class DataSelection(html.Div):
